requirements.txt
# Core ML & Data Science
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Web Framework & API
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Interface & Web
streamlit>=1.28.0
Pillow>=10.0.0

# AI & LLM
openai>=1.3.0
transformers>=4.35.0

# Configuration & Utilities
python-dotenv>=1.0.0
requests>=2.31.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0